\name{trim.data}
\alias{trim.data}
\docType{data}

\title{
Data for Differet Mean Functions
}

\description{
Randomly generated data to showcase different mean functions
}

\usage{data(trim.data)}

\format{
  The format is:
 num [1:10] 1.09 33.79 3.56 5.3 14.19 ...
}

\details{
Data were generated by: rnorm(n = 10, mean = 5, sd = 25)
}

\examples{
data(trim.data)
}

\keyword{datasets}
